Added new Scenes: -WaitingGame -PlayerSelection -CreateAccount -MapSelection -GameMenu
Updated the game manager, scene manager, UI manager, and other classes to fix a minor problem with the data struct.
Updated some scenes with minor changes.
Now the background will be changed when changing the scene and it is not needed to add that code into the finish of every scene.
Moved some includes of the scenes to the scene file to make it easy and faster to create new scenes.